Rework translate_execve() and its support for script/ELF interpreter.
Even if "./script.sh" is equivalent to "/tmp/script.sh" from the
interpreter's point of view, the value of argv[0] in "script.sh" is
not the same.  As a consequence the CoreUtils' configuration script
doesn't work anymore and its test-suite reports errors due to an
unexpected value for $0.  The wrong behaviour was highlighted by
commit 9b82227a.