Be more picky on argv[0] when PRoot uses itself as "launcher".
Technically PRoot uses "/proc/self/exe" as an initial "executor" to
catch the first call to execve(2).  Previously the comparison with
argv[0] didn't allow the non-interactive launch of programs named
".*proot" (technically it was the same effect as calling proot again).
Now only programs named "proot" suffer from this problem, but it's an
acceptable behavior IMHO.