Add support for "asymmetric" path mirrors + sanitize the mirror location.
The command-line option for mirrors was extended to support the syntax
"-m <p1>:<p2>" where <p1> is the location of the mirror within the
alternate rootfs and <p2> is the path to the real directory/file.  For
instance you can now mirror the whole host rootfs in the directory
"/hostfs" within the alternate rootfs that way:

    proot -m /:/hostfs ...

Furthermore, the mirror location is now sanitized (i.e. canonicolized
at the startup) to ensure we can use strncmp() to compare paths
safely.

Note the "asymmetric" support for path mirroring is not yet complete
due to a bug in the detranslation of readlink (and maybe getpwd too).