Rewrite interpreter handlers + introduce naming convention for "path" variables
There was a bug in the check used to ensure the interpreter does not
use an interpreter in its turn: the check was made against the
untranslated (user provided) path to the interpreter!  Technically
PRoot was testing programs out of the "fake" rootfs.

To avoid this kind of bug/thinko I introduced a naming convention for
variables that hold paths, from now on these variables are prefixed
with:

    * "u_", for paths that come from the user-space

    * "t_", for translated paths (canonicalized + rootfs)

    * "d_", for de-translated paths (canonicalized / rootfs)