Create the path up to the mirror location to cheat "walking" programs.
Technically you can create a mirror at any location within the
alternate rootfs however it could bug programs if it is located inside
an non-existant path, for instance "/home/cedric/perl" where only the
"/home" part exits in the alternate rootfs.  Since the entry "cedric"
does not exist in the dierctory "/home", programs that walk along
directories consider "/home/cedric/perl" as an invalid path.

Note that even the final component of the "dummy" path is a directory,
actually its type doesn't matter since only the entry in the parent
directory is important to cheat "walkers".