Add option "-k <string>" to make uname(2) report the specified kernel release.
Technically the GNU C library relies on syscalls provided by the
kernel that's why it performs a sanity check at each application
start-up to verify whether the considered kernel is known to be
compatible.  If you are running a GNU C library that expects a kernel
more recent than your actual one, you will get the following error:

    FATAL: kernel too old

This is where the option "-k <string>" comes in handy: it allows you
to cheat this sanity check by faking the release number of your
kernel.  This option has to be used with care since it does not
improve the compatibility, it just allow to by-pass the sanity check.