Fix the ELF interpreter extraction.
When a PT_INTERP entry is found, the position within the file is
changed by "lseek(fd, segment_offset, SEEK_SET)", as a consequence the
next "read(fd, &program_header, elf_phentsize)" is made at this new
location.  First it's useless to read the next entries once the
PT_INTERP one is found and then there was a small probability that
program_header.p_type was equal to PT_INTERP, so the first extracted
ELF interpreter was replaced by something completely irrelevant.