Use the same prerequisites for the targets %.d and %.o.
It ensures the prerequisites of %.o are now recomputed when one of its
prerequisites has changed.  For instance:

    /* header1.h */
    #include "header2.h"
    #ifdef FOO
    #  include "foo.h"
    #else
    #  include "bar.h"
    #endif

    /* header2.h */
    #define FOO

    /* main.c */
    #include "header1.h"

If ``#define FOO`` is removed from "header2.h" then "main.d" is
changed from:

    test.o test.d: test.c header1.h header2.h foo.h

to:

    test.o test.d: test.c header1.h header2.h bar.h

As you can see, "test.o" now really depends on "bar.h" instead of
"foo.h".  That was not the case with the previous rules.