Fix *partially* support for readlink(2) when mirror pathes are in use.
Prior this patch, any symbolic link -- that points to an absolute path
which prefix is equal to the host-side of any mirror path -- was
bugged.  For instance, the command "proot -m /bin:/host $ROOTFS
/usr/bin/readlink /usr/bin/ps" returned "/host" instead of "/bin/ps".

This fix is *partial* because this support should be rewritten (see
"doc/roadmap.txt").  Moreover I assume in this patch that
"sanity_check" means "not in the context of readlink(2)".