Choose a heuristic regarding bindings and the detranslation of symbolic links.
The detranslation of a symlink should follows bindings only if the
symlink itself belongs to a binding.

For instance, the symbolic link ``/bin/sh -> /bin/bash`` exists both
in the guest rootfs and in the host rootfs, and the host directory
``/bin`` is bound to the guest location ``/foo`` (``-b /bin:/foo``),
then how to resolve the symlinks ``/bin/sh`` and ``/foo/sh``?

With the current heuristic, the results are:

    /bin/sh -> /bin/bash (don't follow bindings)
    /foo/sh -> /foo/bash (follow bindings)

Prior this patch the results were (depending on the intermediate
revision):

    /bin/sh -> /bin/bash
    /foo/sh -> /bin/bash

or:

    /bin/sh -> /foo/bash
    /foo/sh -> /foo/bash

This problem was somewhat already exposed by test-d1da0d8d and
test-d2175fc3.