Re-introduce the "ignore ELF interpreter" option, hidden though.
This option was initially a workaround for what I called the "look
ma', no [heap]" kernel bug:

    $ /lib64/ld-linux-x86-64.so.2 /usr/bin/grep heap /proc/self/maps
    7f56d3bf0000-7f56d3c32000 rw-p 00000000 00:00 0               [heap]

    $ ulimit -s unlimited
    $ /lib64/ld-linux-x86-64.so.2 /usr/bin/grep heap /proc/self/maps

I've removed this option because I'd found a more generic workaround:
set the hard stack limit to a not-so-high value.

It appeared later that this option has an interesting property: it
ensures argv[0] is kept unchanged since there's no need to specify to
the ELF interpreter the [uncanonicalized] path to the executable
(unlike QEMU, ld-linux.so hasn't an option to force argv[0]).  This
behaviour is useful for validation purpose (the Coreutils testsuite
compares hard-coded argv[0] strings) and for some plugins.