Remove specific support for "restart_syscall".
This specific support was written a long time ago to be able to attach
a process on-the-fly.  This feature isn't supported anymore and I wonder
what were the negative effects of this support, for instance, I saw many
"forbidden syscall" messages on Open Build Service computers.  Sadly I'm
not able to produce a test case...

Change-Id: If98986a0fe5c6eb381adfa11267e6d9f707e5230
Signed-off-by: Cédric VINCENT <cedric.vincent@st.com>