Return correct errno if a non-final path component isn't a directory/symlink.
The error is "No such file or directory" if this component doesn't
exist, otherwise the error is "Not a directory".  This sanity check is
bypassed if the canonicalization is made in the name of PRoot itself
(pid == 0) since this would return a spurious error during the
creation of "deep" guest binding paths.

Signed-off-by: Cédric VINCENT <cedric.vincent@st.com>