Add a new command-line interface thanks to the new -r/--rootfs option.
This new command-line interface is more intuitive and it permits the
use of shell aliases:

    $ alias armedslack='proot -Q qemu-arm -r /path/to/armedslack'
    $ armedslack -v 1 -b ~/arm_cpuinfo:/proc/cpuinfo

That wasn't possible previously because no option could be specified
after the guest rootfs.  Thanks to Laurent ALFONSI for convinced me to
implement such a new interface.  Note that the old command-line
interface isn't documented anymore but it is still supported.

Change-Id: I13278436ff9b78020c07c629f5c5213853f7ad97