Added support for architectures with no misalignment support.
The functions in mem.c to read from tracee and write to tracee
memory use word_t access size for copying memory buffers
through the ptrace interface.
This may cause misaligned accesses on architecture where
read/write of word_t size from a misaligned pointer is not
supported. This is the case on sh4 architecture.
The change uses memory accesor function supporting misaligned
access.
To be noted, the ptrace interface seems to allow misaligned
read/write to tracee process through PEEK/POKE, thus no change
was necessary there.

For instance in exit.c a call to copy_from_tracee() is done
for getting the result utsname buffer from the tracee into a
local utsname buffer. This buffer is a char-aligned structure
which may be not aligned on a word boundary.

Reproduce it on sh4 for instance with:
$ make CFLAGS=-O2 -C src
$ src/proot -k 2.6.18 / uname -m
$ dmesg | tail
Unaligned userspace access in "proot" pid=xxxx pc=xxxx ins=0x07ee

Change-Id: I9302781f1fd6bd5eaf13eca27aab3166acddea5f
Signed-off-by: Christophe Guillon <christophe.guillon@st.com>