Emulate syscalls available in the kernel specified by -k, if necessary.
PRoot can now emulate some of the syscalls that are available in the
kernel release specified by -k but that are missing in the host
kernel.  This allows the execution of guest programs expecting a
kernel newer than the actual one, if you encountered the famous
"FATAL: kernel too old" or "qemu: Unsupported syscall" messages.

Change-Id: Id955838ca4c6bf931736b5a8a91f2fae39278c9c