Simplify the operations on tracees' registers by caching them.
The initial purpose of this change was to reduce the number of calls
to ptrace(2) in order to improve the tracing performance.  Actually
the speed-up is just about a couple of percents, however I keep it
because it improves the maintainability of the code a lot by removing
the requirement for error checking when reading/changing a register.