Add support to use Valgrind against the mixed-mode + fix ldso_env_passthru().
Valgrind explicitely replaces "execve(FOREIGN_BINARY)" by
"execve(/bin/sh FOREIGN_BINARY)" (ref: do_exec() in
coregrind/m_ume/main.c) and relies on LD_PRELOAD.  As a consequence,
the mixed-mode validation has to be Valgrind-aware and it has to force
PRoot to consider "/bin/sh" as a foreign binary.

Thanks to this new support, an invalid read in ldso_env_passthru() was
detected: one can't use a pointer to an item once this latter was
over-written: write_item() free the previous item and allocate a new
one.

Change-Id: If1df14427849b996b9dafc2187517e15f28f96e9