Fix binding support when the mount point has no parent directory.
PRoot can now mount/bind files to any location in the guest rootfs,
even if the mount point has no parent directory and the user can't
create it.  Prior this patch, that would created kinda blackhole in
the filesystem hierarchy that might bug some programs like "cpio" or
"rpm".

For example, prior this patch:

   $ proot -b /etc/motd:/black/holes/and/revelations
   proot warning: can't create the guest path (binding) ...
   proot info: started

   $ find /black
   find: `/black: No such file or directory

   $ cat /black/holes/and/revelations
   Time has come to make things right -- Matthew Bellamy

And now:

   $ proot -b /etc/motd:/black/holes/and/revelations
   proot info: started

   $ find /black
   /black
   /black/holes
   /black/holes/and
   /black/holes/and/revelations

   $ cat /black/holes/and/revelations
   Time has come to make things right -- Matthew Bellamy

Change-Id: I2db2e133150d6390b0b2d71c7077b9594636cb05