Rely on the canonicalization loop to create the "glue" rootfs.
As a reminder: the "glue" rootfs is used to allow operations on paths
that do not exist in the guest rootfs but that were specified as the
guest part of a binding. For example, assuming the host path "/opt" is
mounted/bound to the guest path "/black/holes/and/revelations", and
assuming this path can't be created in the guest rootfs
(eg. permission denied), then it is created in a temporary rootfs and
all these paths are glued that way:

    $GUEST/black/ --> $GLUE/black/
                                ./holes
                                ./holes/and
                                ./holes/and/revelations --> $HOST/opt/

Change-Id: Ia0d3cd3b40e8845b3cb3e0081cf76ef3e5951d48