Fix the creation of the glue for relative bindings, eg. ``-b .``
The glue wasn't created for relative bindings because the sanitization
started from getcwd() instead of "/".  For example, assuming "/home"
isn't writable in the guest rootfs and it doesn't contain "cedric",
when running ``~$ proot -b . [...]`` there was no opportunity to glue
"cedric" to "/home" since the sanitization started from
"/home/cedric".

Change-Id: Ic8f6495178491d463cf1e984b488a17eb6a31a81