Add a workaround when QEMU is compiled as a Position Independent Executable.
There's a bug when using the ELF interpreter as a loader (as PRoot
does) on PIE programs that uses constructors (typically QEMU v1.1+).
In this case, constructors are called twice as you can see on the
test below:

     $ cat test.c
     static void __attribute__((constructor)) init(void) { puts("OK"); }
     int main() { return 0; }

     $ gcc -fPIC -pie test.c -o test
     $ ./test
     OK

     $ /lib64/ld-linux-x86-64.so.2 ./test
     OK
     OK

Actually, PRoot doesn't have to use the ELF interpreter as a loader
if QEMU isn't nested.  When QEMU is nested (sub reconfiguration), the
user has to use either a version of QEMU prior v1.1 or a QEMU
compiled with the --disable-pie option.

For information, the error reported by QEMU is something like that:

    $ /lib64/ld-linux-x86-64.so.2 ./qemu-arm
    Registering `arm-cpu' which already exists
    Aborted

Change-Id: I64c849c0a663ea9d2b5803d33cce7f89c3f35000