fake_id0: change file permissions on the fly
Some programms are relaying on the fact that root do not need any permissions
on files and directories to read/write/execute/browse it.
This commit change file and directory permissions on the fly to allows
processes owned by non-privileges users to access such files and directories.

For instance:
$ ls -ld /tmp/test
d--------- /tmp/bla/
$ cd /tmp/test
[Now succeed with this patch]

Change-Id: I379f8404a07882b6fe45bad4b6db09b9390b92c4