Add a NULL terminating byte to readlink*(2) output, when possible.
The output buffer of readlink(2) might be initialized with zeros but
it was updated with the kernel result, and then with the detranslated
result.  This later might be shorter than the former, so it's safier
to add a NULL terminating byte when possible.  This problem was
exposed by IDA Demo 6.3.