Extend syntax of -b option to explicitly not dereference the guest location.
If the guest location is a symbolic link, it is dereferenced to ensure
the new content is accessible through all the symbolic links that
point to the overlaid content.  In most cases this default behavior
shouldn't be a problem, although it is possible to explicitly not
dereference the guest location by prepending it the ``!`` character:
``-b *host_path*:!*guest_location*``.

For example, on Debian and Ubuntu where ``/bin/sh`` is a symbolic link
to ``/bin/dash``.  If ``/bin/bash`` is bound to ``/bin/sh``, the
binding location will be actually ``/bin/dash``:

    proot -b /bin/bash:/bin/sh

    $ md5sum /bin/sh
    089ed56cd74e63f461bef0fdfc2d159a  /bin/sh
    $ md5sum /bin/bash
    089ed56cd74e63f461bef0fdfc2d159a  /bin/bash
    $ md5sum /bin/dash
    089ed56cd74e63f461bef0fdfc2d159a  /bin/dash

It is now possible to strictly bind ``/bin/bash`` over ``/bin/sh`` --
without dereferencing it -- by specifying the ``!`` character::

    proot -b '/bin/bash:!/bin/sh'

    $ md5sum /bin/sh
    089ed56cd74e63f461bef0fdfc2d159a  /bin/sh
    $ md5sum /bin/bash
    089ed56cd74e63f461bef0fdfc2d159a  /bin/bash
    $ md5sum /bin/dash
    c229085928dc19e8d9bd29fe88268504  /bin/dash

Reported-by: Benjamin BARATTE <benjamin.baratte@st.com>