Don't abort when a "spurious" sysexit event from a signaled tracee is reported.
When a tracee gets signaled during its "sysenter" stage, all ptrace
requests return "No such process", so PRoot removes its entry from the
list of tracees.  However, the kernel does report a "sysexit" event,
and there are no way of checking if this is a spurious event since all
ptrace requests succeed:

    1. ptrace event: sysenter (SIGTRAP | 0x80)

    2. the current tracee gets signaled

    3. ptrace request -> "No such process"

    4. ptrace event: sysexit (SIGTRAP | 0x80)

    5. ptrace request -> OK

    6. ptrace event: signaled

The solution is to check whether the tracee structure is sane.