Disable heap emulation on x86 because of a bug that appears on AMD CPUs only.
This bug was exposed by the OBS validation, it could be avoided if the
environment variable "LD_HWCAP_MASK" is set to 0.  I have no ideas
where this bug comes from yet (PRoot, glibc, virtualization, ...).  By
the way, the heap emulation is not required on x86/32-bit.

Thanks to Christophe MONAT <christophe.monat@st.com> for the hint.