Make ptrace/seccomp even more portable on Ubuntu.
The seccomp-filter feature was merged into the Ubuntu kernel before
being accepted upstream (resp. Linux 3.2 & Linux 3.5).  At that time,
PTRACE_EVENT_SECCOMP was defined to 8 on Ubuntu, but it was defined to
7 by upstream.  Today, we have to support both values at the same time
because PRoot might be built with the old value but ran on a kernel
that supports only the new one, or vice versa.

Reported-by:  Maxence Dalmais (Maxired) maxired@gmail.com