Add new "-i uid:gid" option to change the current user and group.
This option makes the current user and group appear as *uid* and
*gid*.  Likewise, files actually owned by the current user and group
appear as if they were owned by *uid* and *gid* instead.  Note that
the ``-0`` option is the same as ``-i 0:0``.

Thanks to this new feature, more tests from the LTP (Linux Test
Project) testsuite now pass when ran through "proot -0":

  Before:

    Total Tests: 991
    Total Failures: 185

  Now:

    Total Tests: 991
    Total Failures: 128

There are no regressions when running this testsuite with "proot -i
$(id -u):$(id -g)".