Re-enable heap emulation support on x86.
This support was historically disabled because it exposes a serious
problem (all tests fail) on http://build.opensuse.org.  After some
tests, this problem appears only on some [AMD processors, hypervisors]
couples.  I was not able to investigate this AMD/hypervisor/heap bug
correctly (no remote shell on build.opensuse.org), and the heap
emulation support is required when ASLR is disabled (as it's now the
case), so I re-enable it.  With luck (hum!), no one will ever hit this
bug.