Make obj-ification in the build-system works whatever $LANG is.
A regular expression is used to know the format and architecture used
by the toolchain, this information is required to link random stuff
(manual, licenses) with the final binary (care or proot).  This
regular expression matches only the text if "LANG=C".

Reported-by: Rémi Duraffort <remi.duraffort@st.com>
Reproduced-with: CARE http://reproducible.io

Yes, CARE is used to reproduce bugs in CARE's build-system :)