Don't emulate suspicious call to brk(2), they might be legit after all.
PRoot might believe that a tracee is calling brk(2) for the first time
with a specific address (this is not supposed to happen the first
time).  It is likely because such tracee is the very first child of
PRoot but the first execve(2) didn't happen yet (so this is not its
first call to brk(2)).  For instance, the installation of seccomp
filters is made after this very first process is traced, and might
call malloc(3) before the first execve(2).

This issue was exposed on a minimal Debian Sid x86_64:

    root@kvm03:~/PRoot# ./src/care -o test.cpio cat
    [...]
    care warning: process 13255 is doing suspicious brk()
    care info: pid 13255: terminated with signal 11
    [...]

Reported-by: Rémi Duraffort <remi.duraffort@st.com>