Change semantic of "HOST_PATH" extension event's second argument.
This second argument ("is_final") used to be true when the current
canonicalized component was the last from the current canonicalized
path.  However, when this last component is a symlink that has to be
dereferenced, another HOST_PATH event with "is_final == true" is
emitted.  In other word, "is_final == true" didn't imply that the
first argument (ie. "host_path") was the result of the
canonicalization.

Now, "is_final == true" implies "host_path" is the result of the
canonicalization, as expected by most extensions ("fake_id0" and
"link2symlink").