Put the heap as close to the BSS as possible.
Some programs -- like dump-emacs -- assume the gap between the end of
the BSS and the start of the heap is relatively small (ie. < 1MB) even
if ALSR is enabled.  Now, the gap between the BSS and the heap is only
"heap_offset" bytes long.  To emulate ADDR_NO_RANDOMIZE personality,
this gap should be removed (not yet supported).