Don't make uthash use talloc, explicitly free memory used by uthash instead.
Each talloc chunk costs 88 bytes on x86_64 (as of talloc 2.1.1).  This
might become significant when allocating a lot of small chunks.  This
issue was exposed by wiom, not by uthash, however one can't reliably
assume how uthash behave in this regard.  It depends on how uthash is
configured, on future changes, ...