fake_id0: Make 32-bit suid binaries raise their faked privileges on PRoot 64-bit.
PR_execve shouldn't be handled in SYSCALL_EXIT_END: the original
version of the register that defines the ABI is overwritten -- on
purpose -- by transfer_load_script().  As a consequence, when
executing a 32-bit program on a 64-bit version of PRoot, get_abi()
returned ABI_2 instead of DEFAULT (remember, the ABI change is
effective only once the syscall has fully returned).