canon: call bindings substitution on '/' component of user path (#100)
Force binding substitution and invocation of HOST_PATH extensions
on the first component of a user path (i.e. '/').

Otherwise for instance the following does not work when the rootfs
is readonly with fake_id0 extension:
$ proot -O -r rootfs touch /empty

This change will scan both '/' and 'empty' path components when
the '/empty' path is canonicalized, formerly, only the 'empty'
component was scanned.