Fix waitpid on zombies
When the ptracer waits on a non-immediate child,
the information is only available in the zombies list and we need to handle those ourselves.

Expands the tracee lookup of the wait enter to include zombies of the tracee to handle this.

The new test mimicing the testing operation from gdb's `linux_check_ptrace_features`.